# calendly-clone-api/docs/en/api.yaml
openapi: 3.0.0
info:
  title: Calendly Clone API
  description: API for scheduling and managing appointments (English documentation)
  version: 1.0.0
servers:
  - url: /
    description: API endpoints
paths:
...
  /appointments:
    post:
      security:
        - BearerAuth: []
      tags:
        - Appointments
      summary: Schedule an appointment
      description: Schedule a new appointment for an event with a user and invitee.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
            example:
              eventId: "evt-123e4567-e89b-12d3-a456-426614174000"
              userId: "user-123e4567-e89b-12d3-a456-426614174000"
              inviteeEmail: "jane.smith@example.com"
              startTime: "2024-01-15T10:00:00Z"
              endTime: "2024-01-15T10:30:00Z"
              status: "scheduled"
      responses:
        '201':
          description: Appointment scheduled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              example:
                id: "apt-123e4567-e89b-12d3-a456-426614174000"
                eventId: "evt-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                inviteeEmail: "jane.smith@example.com"
                startTime: "2024-01-15T10:00:00Z"
                endTime: "2024-01-15T10:30:00Z"
                status: "scheduled"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid input"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      security:
        - BearerAuth: []
      tags:
        - Appointments
      summary: List all appointments
      description: Retrieve a list of all appointments.
      responses:
        '200':
          description: List of appointments retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
              example:
                - id: "apt-123e4567-e89b-12d3-a456-426614174000"
                  eventId: "evt-123e4567-e89b-12d3-a456-426614174000"
                  userId: "user-123e4567-e89b-12d3-a456-426614174000"
                  inviteeEmail: "jane.smith@example.com"
                  startTime: "2024-01-15T10:00:00Z"
                  endTime: "2024-01-15T10:30:00Z"
                  status: "scheduled"
                - id: "apt-223e4567-e89b-12d3-a456-426614174001"
                  eventId: "evt-223e4567-e89b-12d3-a456-426614174001"
                  userId: "user-123e4567-e89b-12d3-a456-426614174000"
                  inviteeEmail: "bob.johnson@example.com"
                  startTime: "2024-01-16T14:00:00Z"
                  endTime: "2024-01-16T15:00:00Z"
                  status: "scheduled"
                - id: "apt-323e4567-e89b-12d3-a456-426614174002"
                  eventId: "evt-323e4567-e89b-12d3-a456-426614174002"
                  userId: "user-123e4567-e89b-12d3-a456-426614174000"
                  inviteeEmail: "alice.white@example.com"
                  startTime: "2024-01-15T15:00:00Z"
                  endTime: "2024-01-15T15:15:00Z"
                  status: "completed"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /appointments/{appointmentId}:
    patch:
      security:
        - BearerAuth: []
      tags:
        - Appointments
      summary: Partially update an appointment
      description: Update specific fields of an appointment by its ID.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the appointment to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentPatch'
            example:
              status: "canceled"
              startTime: "2024-01-15T11:00:00Z"
              endTime: "2024-01-15T11:30:00Z"
      responses:
        '200':
          description: Appointment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              example:
                id: "apt-123e4567-e89b-12d3-a456-426614174000"
                eventId: "evt-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                inviteeEmail: "jane.smith@example.com"
                startTime: "2024-01-15T11:00:00Z"
                endTime: "2024-01-15T11:30:00Z"
                status: "canceled"
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Invalid appointment time or status"
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Appointment with ID apt-123e4567 not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      security:
        - BearerAuth: []
      tags:
        - Appointments
      summary: Delete an appointment
      description: Delete a specific appointment by its ID.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the appointment to delete.
      responses:
        '204':
          description: Appointment deleted successfully
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Resource not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      security:
        - BearerAuth: []
      tags:
        - Appointments
      summary: Get appointment details
      description: Retrieve details of a specific appointment by its ID.
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
          description: The ID of the appointment to retrieve.
      responses:
        '200':
          description: Appointment details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
              example:
                id: "apt-123e4567-e89b-12d3-a456-426614174000"
                eventId: "evt-123e4567-e89b-12d3-a456-426614174000"
                userId: "user-123e4567-e89b-12d3-a456-426614174000"
                inviteeEmail: "jane.smith@example.com"
                startTime: "2024-01-15T10:00:00Z"
                endTime: "2024-01-15T10:30:00Z"
                status: "scheduled"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Authentication required"
        '404':
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: "Appointment with ID apt-123e4567 not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Appointment:
      type: object
      properties:
        id:
          type: string
        eventId:
          type: string
        userId:
          type: string
        inviteeEmail:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [scheduled, canceled, completed]
      required:
        - eventId
        - userId
        - inviteeEmail
        - startTime
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: "Bad Request"
      required:
        - message
...
    AppointmentPatch:
      type: object
      properties:
        eventId:
          type: string
        userId:
          type: string
        inviteeEmail:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      minProperties: 1  # At least one field must be provided
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Token